Revert selector labels from commons
The initial commit that changed this mistakenly assumed all selectors where the same. This commit reverts that. As selector labels are immutable they cannot be changed.

However they will be in a future release to the desired selector labels. A guide on how to do the migration will be included when the time comes.